90% complete 🤩✨

+++ app/src/main/java/com/entsh104/highking/ui/cust/profile/ProfileFragment.kt
@@ -1,7 +1,9 @@
 package com.entsh104.highking.ui.cust.profile
 
 import UserRepository
+import android.app.AlertDialog
 import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
 import android.util.Log
 import android.view.LayoutInflater
@@ -9,8 +11,12 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.Toast
+import androidx.activity.result.PickVisualMediaRequest
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.fragment.app.Fragment
+import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
+import androidx.lifecycle.repeatOnLifecycle
 import androidx.navigation.fragment.findNavController
 import com.bumptech.glide.Glide
 import com.entsh104.highking.R
@@ -18,8 +24,7 @@ import com.entsh104.highking.data.source.local.SharedPreferencesManager
 import com.entsh104.highking.data.source.remote.RetrofitClient
 import com.entsh104.highking.databinding.FragmentCustProfileBinding
 import com.entsh104.highking.ui.auth.AuthActivity
-import com.entsh104.highking.ui.cust.CustActivity
-import com.entsh104.highking.ui.util.NavOptionsUtil
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 
 class ProfileFragment : Fragment() {
@@ -49,19 +54,29 @@ class ProfileFragment : Fragment() {
 
         binding.btnLogout.setOnClickListener {
             val token = prefs.getToken().toString()
-            lifecycleScope.launch {
-                val responseLogout = userRepository.logoutUser(token)
-                if (responseLogout.isSuccess) {
-                    Toast.makeText(requireContext(), "Logout Berhasil", Toast.LENGTH_SHORT).show()
-                    prefs.clear()
-                    val intent = Intent(activity, AuthActivity::class.java)
-                    startActivity(intent)
-                    activity?.finish()
-                } else {
-                    Toast.makeText(requireContext(), "Logout failed", Toast.LENGTH_SHORT).show()
+            viewLifecycleOwner.lifecycleScope.launch {
+                lifecycle.repeatOnLifecycle(Lifecycle.State.STARTED) {
+                    delay(500)
+                    if (viewLifecycleOwner.lifecycle.currentState.isAtLeast(Lifecycle.State.STARTED)) {
+                        val responseLogout = userRepository.logoutUser(token)
+                        if (responseLogout.isSuccess) {
+                            Toast.makeText(requireContext(), "Logout Berhasil", Toast.LENGTH_SHORT)
+                                .show()
+                            prefs.clear()
+                            val intent = Intent(activity, AuthActivity::class.java)
+                            startActivity(intent)
+                            activity?.finish()
+                        } else {
+                            Toast.makeText(requireContext(), "Logout failed", Toast.LENGTH_SHORT)
+                                .show()
+                        }
+                    }
                 }
-
             }
+        }
+
+        binding.buttonEditPhoto.setOnClickListener {
+            startGallery()
         }
 
         binding.buttonEditProfile.setOnClickListener {
@@ -85,29 +100,92 @@ class ProfileFragment : Fragment() {
         fetchUserProfile()
     }
 
-    private fun fetchUserProfile() {
-        lifecycleScope.launch {
-            binding.progressBar.visibility = View.VISIBLE
-            binding.scrollViewContent.visibility = View.GONE
-
-            val result = userRepository.getCurrentUser()
-            if (result.isSuccess) {
-                val user = result.getOrNull()
-                user?.let {
-                    Log.d("ProfileFragment", "User: $it")
-                    binding.textViewUsername.text = "@" + it.username
-                    binding.textViewName.text = it.username
-                    binding.textViewEmail.text = it.email
-                    binding.textViewPhone.text = it.phone_number
-                    Glide.with(this@ProfileFragment).load("${it.image_url}")
-                        .into(binding.imageViewProfile)
-                    binding.progressBar.visibility = View.GONE
-                    binding.scrollViewContent.visibility = View.VISIBLE
+    private fun showConfirmationDialog(imageUri: Uri) {
+        val builder = AlertDialog.Builder(requireContext())
+        builder.setTitle("Confirm Photo")
+        builder.setMessage("Are you sure you want to upload this photo?")
+        builder.setPositiveButton("Yes") { dialog, which ->
+            viewLifecycleOwner.lifecycleScope.launch {
+                lifecycle.repeatOnLifecycle(Lifecycle.State.STARTED) {
+                    delay(500)
+                    if (viewLifecycleOwner.lifecycle.currentState.isAtLeast(Lifecycle.State.STARTED)) {
+                        val response = userRepository.uploadPhoto(
+                            requireContext().contentResolver,
+                            requireContext().cacheDir,
+                            imageUri
+                        )
+                        if (response.isSuccess) {
+                            Toast.makeText(
+                                requireContext(),
+                                "Photo uploaded successfully",
+                                Toast.LENGTH_SHORT
+                            ).show()
+                            fetchUserProfile()
+                        } else {
+                            Toast.makeText(
+                                requireContext(),
+                                "Failed to upload photo",
+                                Toast.LENGTH_SHORT
+                            ).show()
+                        }
+                    }
                 }
-            } else {
-                Toast.makeText(requireContext(), "Failed to load profile", Toast.LENGTH_SHORT)
-                    .show()
-                binding.progressBar.visibility = View.GONE
+            }
+        }
+        builder.setNegativeButton("No") { dialog, which ->
+            dialog.dismiss()
+        }
+        builder.show()
+    }
+
+
+    private val launcherGallery = registerForActivityResult(ActivityResultContracts.PickVisualMedia()) { uri: Uri? ->
+        uri?.let {
+            showConfirmationDialog(it)
+        }
+    }
+
+    private fun startGallery() {
+        launcherGallery.launch(
+            PickVisualMediaRequest(
+                ActivityResultContracts.PickVisualMedia.ImageOnly
+            )
+        )
+    }
+
+
+    private fun fetchUserProfile() {
+        viewLifecycleOwner.lifecycleScope.launch {
+            lifecycle.repeatOnLifecycle(Lifecycle.State.STARTED) {
+                delay(500)
+                if (viewLifecycleOwner.lifecycle.currentState.isAtLeast(Lifecycle.State.STARTED)) {
+                    binding.progressBar.visibility = View.VISIBLE
+                    binding.scrollViewContent.visibility = View.GONE
+
+                    val result = userRepository.getCurrentUser()
+                    if (result.isSuccess) {
+                        val user = result.getOrNull()
+                        user?.let {
+                            Log.d("ProfileFragment", "User: $it")
+                            binding.textViewUsername.text = "@" + it.username
+                            binding.textViewName.text = it.username
+                            binding.textViewEmail.text = it.email
+                            binding.textViewPhone.text = it.phone_number
+                            Glide.with(this@ProfileFragment).load("${it.image_url}")
+                                .into(binding.imageViewProfile)
+                            binding.progressBar.visibility = View.GONE
+                            binding.scrollViewContent.visibility = View.VISIBLE
+                        }
+                    } else {
+                        Toast.makeText(
+                            requireContext(),
+                            "Failed to load profile",
+                            Toast.LENGTH_SHORT
+                        )
+                            .show()
+                        binding.progressBar.visibility = View.GONE
+                    }
+                }
             }
         }
     }
